[WFLY-19198] Move the OIDC scope related tests back to the elytron-oidc-client testsuite: add tests to wildfly testsuite
