--- testsuite/integration/manualmode/src/test/java/org/wildfly/test/manual/elytron/oidc/OidcBaseTest.java
@@ -1,302 +1,0 @@
-/*
- * Copyright The WildFly Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.wildfly.test.manual.elytron.oidc;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assume.assumeTrue;
-import static org.wildfly.security.http.oidc.Oidc.OIDC_SCOPE;
-
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
-import java.util.LinkedList;
-import java.util.List;
-
-import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
-import org.apache.http.client.CookieStore;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.client.utils.HttpClientUtils;
-import org.apache.http.impl.client.BasicCookieStore;
-import org.apache.http.impl.client.BasicResponseHandler;
-import org.apache.http.impl.client.LaxRedirectStrategy;
-import org.apache.http.message.BasicNameValuePair;
-import org.apache.http.protocol.BasicHttpContext;
-import org.apache.http.protocol.HttpContext;
-import org.jboss.arquillian.container.test.api.ContainerController;
-import org.jboss.arquillian.container.test.api.OperateOnDeployment;
-import org.jboss.arquillian.junit.InSequence;
-import org.jboss.as.arquillian.api.ServerSetupTask;
-import org.jboss.as.arquillian.container.ManagementClient;
-import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.as.test.http.util.TestHttpClientUtils;
-import org.jboss.as.test.integration.security.common.servlets.SimpleServlet;
-import org.jboss.as.test.shared.TestSuiteEnvironment;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.keycloak.representations.idm.RealmRepresentation;
-import org.testcontainers.DockerClientFactory;
-import org.wildfly.security.jose.util.JsonSerialization;
-
-import io.restassured.RestAssured;
-import org.wildfly.test.manual.elytron.oidc.subsystem.SimpleServletWithScope;
-
-/**
- * Tests for the OpenID Connect authentication mechanism.
- *
- * @author <a href="mailto:prpaul@redhat.com">Prarthona Paul</a>
- */
-public abstract class OidcBaseTest {
-
-    public static final String CLIENT_SECRET = "secret";
-    // TODO migrate this to testsuite/integration/basic and drop the ContainerController aspect used in manualmode
-    public static final String CONTAINER = "stability-preview";
-    public static final String OIDC_WITHOUT_SUBSYSTEM_CONFIG_WEB_XML = "web.xml";
-    public static KeycloakContainer KEYCLOAK_CONTAINER;
-    public static final String TEST_REALM = "WildFly";
-    private static final String KEYCLOAK_USERNAME = "username";
-    private static final String KEYCLOAK_PASSWORD = "password";
-    public static final int CLIENT_PORT = TestSuiteEnvironment.getHttpPort();
-    public static final String CLIENT_HOST_NAME = TestSuiteEnvironment.getHttpAddress();
-    public static final String OPENID_SCOPE_APP = "OpenIDScopeApp";
-    public static final String INVALID_SCOPE_APP = "InvalidScopeApp";
-    public static final String SINGLE_SCOPE_APP = "SingleScopeApp";
-    public static final String MULTIPLE_SCOPE_APP = "MultipleScopeApp";
-
-    @ArquillianResource
-    private static ContainerController containerController;
-    private static boolean isDockerAvailable() {
-        try {
-            DockerClientFactory.instance().client();
-            return true;
-        } catch (Throwable ex) {
-            return false;
-        }
-    }
-
-    public static void sendRealmCreationRequest(RealmRepresentation realm) {
-        try {
-            String adminAccessToken = KeycloakConfiguration.getAdminAccessToken(KEYCLOAK_CONTAINER.getAuthServerUrl());
-            assertNotNull(adminAccessToken);
-            RestAssured
-                    .given()
-                    .auth().oauth2(adminAccessToken)
-                    .contentType("application/json")
-                    .body(JsonSerialization.writeValueAsBytes(realm))
-                    .when()
-                    .post(KEYCLOAK_CONTAINER.getAuthServerUrl() + "/admin/realms").then()
-                    .statusCode(201);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @BeforeClass
-    public static void checkDockerAvailability() {
-        assumeTrue("Docker isn't available, OIDC tests will be skipped", isDockerAvailable());
-    }
-
-    @Test
-    @InSequence(-1)
-    public void startContainer() throws Exception {
-        containerController.start(CONTAINER);
-    }
-
-    /**
-     * Tests that use different scope values to request access to claims values.
-     */
-
-    @Test
-    @OperateOnDeployment(OPENID_SCOPE_APP)
-    public void testOpenIDScope() throws Exception {
-        String expectedScope = OIDC_SCOPE;
-        loginToApp(KeycloakConfiguration.ALICE, KeycloakConfiguration.ALICE_PASSWORD, HttpURLConnection.HTTP_OK, SimpleServlet.RESPONSE_BODY, true,
-                new URL("http", TestSuiteEnvironment.getHttpAddress(), TestSuiteEnvironment.getHttpPort(),
-                        "/" + OPENID_SCOPE_APP + SimpleServletWithScope.SERVLET_PATH).toURI(), expectedScope, false);
-    }
-
-    @Test
-    @OperateOnDeployment(SINGLE_SCOPE_APP)
-    public void testSingleScope() throws Exception {
-        String expectedScope = OIDC_SCOPE + "+profile";
-        loginToApp(KeycloakConfiguration.ALICE, KeycloakConfiguration.ALICE_PASSWORD, HttpURLConnection.HTTP_OK, SimpleServlet.RESPONSE_BODY, true,
-                new URL("http", TestSuiteEnvironment.getHttpAddress(), TestSuiteEnvironment.getHttpPort(),
-                        "/" + SINGLE_SCOPE_APP + SimpleServletWithScope.SERVLET_PATH).toURI(), expectedScope, false);
-    }
-
-    @Test
-    @OperateOnDeployment(MULTIPLE_SCOPE_APP)
-    public void testMultipleScope() throws Exception {
-        String expectedScope = OIDC_SCOPE + "+phone+profile+microprofile-jwt+email";
-        loginToApp(KeycloakConfiguration.ALICE, KeycloakConfiguration.ALICE_PASSWORD, HttpURLConnection.HTTP_OK, SimpleServlet.RESPONSE_BODY, true,
-                new URL("http", TestSuiteEnvironment.getHttpAddress(), TestSuiteEnvironment.getHttpPort(),
-                        "/" + MULTIPLE_SCOPE_APP + SimpleServletWithScope.SERVLET_PATH).toURI(), expectedScope, false);
-    }
-
-    @Test
-    @OperateOnDeployment(INVALID_SCOPE_APP)
-    public void testInvalidScope() throws Exception {
-        String expectedScope = OIDC_SCOPE + "+INVALID_SCOPE";
-        loginToApp(KeycloakConfiguration.ALICE, KeycloakConfiguration.ALICE_PASSWORD, HttpURLConnection.HTTP_OK, SimpleServlet.RESPONSE_BODY, false,
-                new URL("http", TestSuiteEnvironment.getHttpAddress(), TestSuiteEnvironment.getHttpPort(),
-                        "/" + INVALID_SCOPE_APP + SimpleServletWithScope.SERVLET_PATH).toURI(), expectedScope, true);
-    }
-
-    public static void loginToApp(String username, String password, int expectedStatusCode, String expectedText, boolean loginToKeycloak, URI requestUri, String expectedScope, boolean checkInvalidScope) throws Exception {
-        CookieStore store = new BasicCookieStore();
-        HttpClient httpClient = TestHttpClientUtils.promiscuousCookieHttpClientBuilder()
-                .setDefaultCookieStore(store)
-                .setRedirectStrategy(new LaxRedirectStrategy())
-                .build();
-        HttpGet getMethod = new HttpGet(requestUri);
-        HttpContext context = new BasicHttpContext();
-        HttpResponse response = httpClient.execute(getMethod, context);
-
-        try {
-            int statusCode = response.getStatusLine().getStatusCode();
-            if (loginToKeycloak) {
-                assertTrue("Expected code == OK but got " + statusCode + " for request=" + requestUri, statusCode == HttpURLConnection.HTTP_OK);
-                Form keycloakLoginForm = new Form(response);
-                HttpResponse afterLoginClickResponse = simulateClickingOnButton(httpClient, keycloakLoginForm, username, password, "Sign In");
-                afterLoginClickResponse.getEntity().getContent();
-                assertEquals(expectedStatusCode, afterLoginClickResponse.getStatusLine().getStatusCode());
-                if (expectedText != null) {
-                    String responseString = new BasicResponseHandler().handleResponse(afterLoginClickResponse);
-                    assertTrue("Unexpected result " + responseString, responseString.contains(expectedText));
-                    if (expectedScope != null) {
-                        assertTrue(context.toString().contains("scope=" + expectedScope));
-                        if (expectedScope.contains("profile")) {
-                            assertTrue(responseString.contains("profile: " + KeycloakConfiguration.ALICE_FIRST_NAME + " " + KeycloakConfiguration.ALICE_LAST_NAME));
-                        }
-                        if (expectedScope.contains("email")) {
-                            assertTrue(responseString.contains("email: " + KeycloakConfiguration.ALICE_EMAIL_VERIFIED));
-                        }
-                        if (expectedScope.contains("microprofile-jwt")) {
-                            assertTrue(responseString.contains("microprofile-jwt: [" + KeycloakConfiguration.JBOSS_ADMIN_ROLE + ", " + KeycloakConfiguration.USER_ROLE + "]"));
-                        }
-                    }
-                }
-            } else if (checkInvalidScope) {
-                assertTrue(context.toString().contains("error_description=Invalid+scopes"));
-                assertTrue("Expected code == BAD REQUEST but got " + statusCode + " for request=" + requestUri, statusCode == HttpURLConnection.HTTP_BAD_REQUEST);
-            } else {
-                assertTrue("Expected code == FORBIDDEN but got " + statusCode + " for request=" + requestUri, statusCode == HttpURLConnection.HTTP_FORBIDDEN);
-            }
-        } finally {
-            HttpClientUtils.closeQuietly(response);
-        }
-    }
-
-    public static class KeycloakSetup implements ServerSetupTask {
-
-        @Override
-        public void setup(ManagementClient managementClient, String containerId) throws Exception {
-            KEYCLOAK_CONTAINER = new KeycloakContainer();
-            KEYCLOAK_CONTAINER.start();
-        }
-
-        public void tearDown(ManagementClient managementClient, String containerId) throws Exception {
-            if (KEYCLOAK_CONTAINER != null) {
-                KEYCLOAK_CONTAINER.stop();
-            }
-        }
-    }
-
-    private static HttpResponse simulateClickingOnButton(HttpClient client, Form form, String username, String password, String buttonValue) throws IOException {
-        final URL url = new URL(form.getAction());
-        final HttpPost request = new HttpPost(url.toString());
-        final List<NameValuePair> params = new LinkedList<>();
-        for (Input input : form.getInputFields()) {
-            if (input.type == Input.Type.HIDDEN ||
-                    (input.type == Input.Type.SUBMIT && input.getValue().equals(buttonValue))) {
-                params.add(new BasicNameValuePair(input.getName(), input.getValue()));
-            } else if (input.getName().equals(KEYCLOAK_USERNAME)) {
-                params.add(new BasicNameValuePair(input.getName(), username));
-            } else if (input.getName().equals(KEYCLOAK_PASSWORD)) {
-                params.add(new BasicNameValuePair(input.getName(), password));
-            }
-        }
-        request.setEntity(new UrlEncodedFormEntity(params, StandardCharsets.UTF_8));
-        return client.execute(request);
-    }
-
-    public static final class Form {
-
-        static final String
-                NAME = "name",
-                VALUE = "value",
-                INPUT = "input",
-                TYPE = "type",
-                ACTION = "action",
-                FORM = "form";
-
-        final HttpResponse response;
-        final String action;
-        final List<Input> inputFields = new LinkedList<>();
-
-        public Form(HttpResponse response) throws IOException {
-            this.response = response;
-            final String responseString = new BasicResponseHandler().handleResponse(response);
-            final Document doc = Jsoup.parse(responseString);
-            final Element form = doc.select(FORM).first();
-            this.action = form.attr(ACTION);
-            for (Element input : form.select(INPUT)) {
-                Input.Type type = null;
-                switch (input.attr(TYPE)) {
-                    case "submit":
-                        type = Input.Type.SUBMIT;
-                        break;
-                    case "hidden":
-                        type = Input.Type.HIDDEN;
-                        break;
-                }
-                inputFields.add(new Input(input.attr(NAME), input.attr(VALUE), type));
-            }
-        }
-
-        public String getAction() {
-            return action;
-        }
-
-        public List<Input> getInputFields() {
-            return inputFields;
-        }
-    }
-
-    private static final class Input {
-
-        final String name, value;
-        final Input.Type type;
-
-        public Input(String name, String value, Input.Type type) {
-            this.name = name;
-            this.value = value;
-            this.type = type;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public String getValue() {
-            return value;
-        }
-
-        public enum Type {
-            HIDDEN, SUBMIT
-        }
-    }
-}
