+++ testsuite/integration/manualmode/src/test/resources/arquillian.xml
@@ -529,19 +529,6 @@
                 <property name="javaHome">${container.java.home}</property>
             </configuration>
         </container>
-        <container qualifier="stability-preview" default="false" mode="manual">
-            <configuration>
-                <property name="jbossHome">${oidc.jboss.home}</property>
-                <property name="javaVmArguments">${server.jvm.args}</property>
-                <property name="serverConfig">${jboss.server.config.file.name:standalone.xml}</property>
-                <property name="jbossArguments">${jboss.args} --stability=preview</property>
-                <property name="allowConnectingToRunningServer">true</property>
-                <property name="managementAddress">${node0:127.0.0.1}</property>
-                <property name="managementPort">${as.managementPort:9990}</property>
-
-                <property name="javaHome">${container.java.home}</property>
-            </configuration>
-        </container>
     </group>
 
 </arquillian>
